requirements.txt
pydantic-ai>=1.56.0
pydantic-settings>=2.12.0
httpx>=0.28.1
aiohttp>=3.9.0
bilibili-api-python>=17.0.0
